\name{is.wrsp}
\alias{is.owrsp}
\alias{is.wrsp}
\title{Reports whether an object has S3 class 'wrsp',
or 'owrsp', as
would be returned by, for example,
\code{\link{well_response}}.}
\usage{
  is.wrsp(Obj)

  is.owrsp(Obj)
}
\arguments{
  \item{Obj}{the object to test}
}
\description{
  Reports whether an object has S3 class 'wrsp', or
  'owrsp', as would be returned by, for example,
  \code{\link{well_response}}.
}
\author{
  A. J. Barbour <andy.barbour@gmail.com>
}
\seealso{
  \code{\link{kitagawa-package}}

  Other utilities: \code{\link{.in0to1}},
  \code{\link{.nullchk}}, \code{\link{omega_norm}},
  \code{\link{sensing_volume}}
}

