\docType{package}
\name{kitagawa-package}
\alias{kitagawa}
\alias{kitagawa-package}
\title{Calculate expected spectral response for high-frequency borehole strain and fluid-pressure measurements}
\description{
  Creates a model of the spectral response of a closed well
  undergoing dynamic, volumetric strains at seismic
  frequencies.  The model may be used to solve for aquifer
  parameters using non-linear inversion techniques.
}
\details{
  Here are some details
}
\author{
  Andrew Barbour <andy.barbour@gmail.com>
}
\references{
  Kitagawa, Y., S. Itaba, N. Matsumoto, and N. Koisumi
  (2011), Frequency characteristics of the response of
  water pressure in a closed well to volumetric strain in
  the high-frequency domain, \emph{J. Geophys. Res.},
  \strong{116}, B08301, doi:10.1029/2010JB007794.

  \url{http://www.agu.org/pubs/crossref/2011/2010JB007794.shtml}
}
\seealso{
  \code{\link{well_response}},
  \code{\link{sensing_volume}}, \code{\link{kitplot}}
}

