\name{omega_norm}
\alias{omega_norm}
\title{Dimensionless frequency from diffusivity and depth}
\usage{
  omega_norm(omega, Diffusiv, z)
}
\arguments{
  \item{omega}{numeric; angular frequency}

  \item{Diffusiv}{numeric; hydraulic diffusivity}

  \item{z}{numeric; depth}
}
\description{
  Dimensionless frequency from diffusivity and depth
}
\seealso{
  \code{\link{open_well_response}},
  \code{\link{kitagawa-package}}

  Other utilities: \code{\link{.in0to1}},
  \code{\link{.nullchk}}, \code{\link{sensing_volume}}
}

